% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_user.R
\name{get_user}
\alias{get_user}
\title{Get information about a user}
\usage{
get_user(user_id = "self")
}
\arguments{
\item{user_id}{The ID of the user to get info for. If none is provided,
information about yourself is returned.}
}
\value{
A list containing information about the user
}
\description{
Get information about a user
}
\examples{
\dontrun{
library(numerous)
u <- get_user()
}
}

