% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/numerous-package.R
\docType{package}
\name{numerous}
\alias{numerous}
\alias{numerous-package}
\title{Interact with the Numerous API}
\description{
\href{http://numerousapp.com}{Numerous} is a service for sharing and
following numbers such as stock prices, temperatures, the number of processes
running on a server, or the number of times you've visited the gym this week.
}
\author{
Brian Connelly
}
\seealso{
\href{https://developer.numerousapp.com}{Numerous API Documentation}
}

